requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.6
django-redis==5.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
Faker==17.0.0
idna==3.4
kombu==5.2.4
kwargs==1.0.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pipreqs==0.4.11
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg2==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing==3.7.4.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
yarg==0.1.9
